refactor: migrate all csr useAPI hook
