+++ pages/index.tsx
@@ -2,10 +2,10 @@ import LandingTemplate from "../components/template/LandingTemplate";
 import { Header } from "../components/organism/landing/Header";
 import React, { useEffect } from "react";
 import { ProjectCard } from "../components/molecule/projectCard/ProjectCard";
-import { useAPINew } from "../api/hook";
+import { useAPILegacy } from "../utils/hook/api";
 
 export default function Home() {
-  const topProjects = useAPINew((api) => api.landing.topProjects);
+  const topProjects = useAPILegacy((api) => api.landing.topProjects);
 
   const projectInfo = topProjects.data?.projectCard;
