+++ pages/project/index.tsx
@@ -7,7 +7,7 @@ import { SingleDropDown as Single } from "../../components/molecule/drop-down/Si
 import { FilterGroupDropDown as Group } from "../../components/molecule/drop-down/FilterGroupDropDown";
 import { ProjectCard } from "../../components/molecule/projectCard/ProjectCard";
 import useInfinityScroll from "../../utils/hook/useInfinityScroll";
-import { useAPINew } from "../../api/hook";
+import { useAPILegacy } from "../../utils/hook/api";
 
 export interface SearchProjectProps {
   projectMetadata: ProjectMeta;
@@ -26,7 +26,7 @@ export default function SearchProject(props: SearchProjectProps) {
   const {
     projectMetadata: { project: meta },
   } = props;
-  const search = useAPINew((api) => api.project.searchProject);
+  const search = useAPILegacy((api) => api.project.searchProject);
   const [projectInfo, setProjectInfo] = useState<ProjectInfo>({ projectCard: [] });
   const [requestInfo, setRequestInfo] = useState<RequestInfo>(initial);
   const [isMoreData, setIsMoreData] = useState<boolean>(true);
